Change Hibernate tests to leverage the new common test suite
Next step is to have better organization of tests, but now we starting to share
the common test cases among multiple implementations.

Also updated the DAOTestSuite to expose the simpleObjectDAO as a protected
property so that subclasses can get to it (required for the HibernateDAOTest
override of findByBogusId (which expects an exception instead of a null value).

Updated the createData method to ensure there was a duplicate value.

Added an ID to the RelatedObject class.
