+++ jeppetto-dao-hibernate/src/test/resources/HibernateDAOTest.spring.xml
@@ -50,26 +50,26 @@
 
     <!-- =================== Data Access Objects =================== -->
 
-    <bean id="simpleObjectDAO" class="org.iternine.jeppetto.dao.DAOBuilder" factory-method="buildDAO">
-        <constructor-arg index="0" value="org.iternine.jeppetto.SimpleObject"/>
-        <constructor-arg index="1" value="org.iternine.jeppetto.SimpleObjectDAO"/>
-        <constructor-arg index="2" value="org.iternine.jeppetto.dao.hibernate.HibernateQueryModelDAO"/>
-        <constructor-arg index="3">
-            <map>
-                <entry key="sessionFactory" value-ref="sessionFactory"/>
-            </map>
-        </constructor-arg>
-    </bean>
-
-    <bean id="transactionalSimpleObjectDAO" class="org.springframework.transaction.interceptor.TransactionProxyFactoryBean">
+    <bean id="simpleObjectDAO" class="org.springframework.transaction.interceptor.TransactionProxyFactoryBean">
         <property name="proxyInterfaces">
             <list>
                 <bean class="java.lang.Class" factory-method="forName">
-                    <constructor-arg value="org.iternine.jeppetto.SimpleObjectDAO"/>
+                    <constructor-arg value="org.iternine.jeppetto.test.SimpleObjectDAO"/>
                 </bean>
             </list>
         </property>
-        <property name="target" ref="simpleObjectDAO"/>
+        <property name="target">
+            <bean class="org.iternine.jeppetto.dao.DAOBuilder" factory-method="buildDAO">
+                <constructor-arg index="0" value="org.iternine.jeppetto.test.SimpleObject"/>
+                <constructor-arg index="1" value="org.iternine.jeppetto.test.SimpleObjectDAO"/>
+                <constructor-arg index="2" value="org.iternine.jeppetto.dao.hibernate.HibernateQueryModelDAO"/>
+                <constructor-arg index="3">
+                    <map>
+                        <entry key="sessionFactory" value-ref="sessionFactory"/>
+                    </map>
+                </constructor-arg>
+            </bean>
+        </property>
         <property name="transactionManager" ref="transactionManager"/>
         <property name="transactionAttributes">
             <props>
