+++ jeppetto-test-functional/src/main/java/org/iternine/jeppetto/test/RelatedObject.java
@@ -23,12 +23,22 @@ public class RelatedObject {
     // Variables - Private
     //-------------------------------------------------------------
 
+    private String id;
     private int relatedIntValue;
 
 
     //-------------------------------------------------------------
     // Methods - Getter/Setter
     //-------------------------------------------------------------
+
+    public String getId() {
+        return id;
+    }
+
+    public void setId(String id) {
+        this.id = id;
+    }
+
 
     public int getRelatedIntValue() {
         return relatedIntValue;
